orders: check changed products from PackageCreated
When the OnEvents function receives a PackageCreated event, it now
updates products in the order. This causes the TableUpdate function to
trigger an event that contains "products" in the "changed" field in the
detail.

This commits also updates the on_events integration tests to check if
events trigger a new OrderModified event.